Makefile: Let users chose to fail on first error
By default the top level makefile was not failing and stopping when
a component could not be built. Instead it was printing a warning
and continuing with whatever it could do.

With this patch it is possible for the user to change this behaviour
by setting the environment variable BSIM_BUILD_FAIL_ASAP.
When set, make will fail as soon as any component fails.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>