libUtilv1: symbols: Handle symbols over 63 chars
Bugfix: When a symbol was over 63 chars, it was not
properly terminated => terminate it.
Increase the size of the symbol name string to 256 chars
to allow for outrageously long named symbols
(e.g. automatically generated names). At this point it
starts to be worth considering allocating them dynamically
in a bit smarter way, as we will be spending around 500KB-4MB
for typical devices just for this table.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>