libRandv2: Add new bs_random_buffer() and deprecate bs_random_uint32()
Add a new API `bs_random_buffer()` which can be used to fill buffers
with pseudorandom numbers.

At the same time, deprecate bs_random_uint32(), as it was buggy.
Note:
We will not fix this old function as doing so would probably cause
old tests to fail due to the change in randomness. We will never remove
this function. Instead we encourage users to switch to
bs_random_buffer().

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>